+++ dsrnascan/__init__.py
@@ -1,5 +1,5 @@
 # dsRNAscan package
 from .dsRNAscan import main
 
-__version__ = '0.5.0'
+__version__ = '0.5.1'
 __all__ = ['main']